AudioClassify/requirements.txt
librosa==0.8.1
tqdm==4.61.2
numpy~=1.19.3
pandas==1.3.3
tensorflow-cpu==2.6.0
sklearn==1.0
jupyterlab==3.2.0